+++ src/app/auth/login/components/index.ts
@@ -1,0 +1,1 @@
+export { default as KakaoButton } from './KakaoButton';
