+++ src/app/place/[id]/components/Header.tsx
