+++ src/app/place/[id]/components/KagongBox.tsx
