+++ src/app/place/[id]/components/Rating.tsx
