+++ src/app/place/[id]/components/TimeInfo.tsx
