+++ src/app/place/[id]/components/index.ts
@@ -1,14 +1,11 @@
 export { default as DefaultInfo } from './DefaultInfo';
 export { default as Header } from './Header';
-export { default as IconFlex } from './IconFlex';
 export { default as KagongBox } from './KagongBox';
 export { default as ReviewBox } from './ReviewBox';
 export { default as Tag } from './Tag';
 export { default as TimeInfo } from './TimeInfo';
 export { default as Tooltip } from './Tooltip';
 export { default as ReviewSheet } from './ReviewSheet';
-export { default as Tabs } from './Tabs';
 export { default as Rating } from './Rating';
 export { default as TimeLogSheet } from './TimeLogSheet';
-export { default as ImageUpload } from './ImageUpload';
 export { default as ShareSheet } from './ShareSheet';
