+++ src/app/place/[id]/gallery/components/index.ts
@@ -1,0 +1,1 @@
+export { default as GalleryModal } from './GalleryModal';
