+++ src/components/Flex/index.ts
@@ -1,0 +1,1 @@
+export { default as IconFlex } from './IconFlex';
