+++ src/components/index.ts
@@ -1,0 +1,7 @@
+export { default as Button } from './Button';
+export { default as BottomSheet } from './BottomSheet';
+export { default as SideMenu } from './SideMenu';
+export { default as ImageUpload } from './ImageUpload';
+export { default as KakaoMap } from './KakaoMap';
+export * from './Flex';
+export * from './Tab';
